druppie/mcp-servers/module-boter-kaas-eieren/requirements.txt
fastmcp==0.4.0
starlette==0.27.0
uvicorn[standard]==0.29.0
pyyaml==6.0.1
